src/content_creator_agency/requirements.txt
agency_swarm
pydantic
python-dotenv
pytrends
tavily-python
nltk
google-api-python-client



